feat(chezmoi): first select dependencies and install in the end
The current implementation requires the user to interact with the
`chezmoi` installation process from start to end. Every time a package
is installed, the user has to wait for it to finish until they can
continue selecting the packages they want to install.

If the entire `build()` part of the packages takes 1 hour in total but
is distributed over small packages, then the user has to interact with
`chezmoi` for that whole time. Instead, it is much more convenient, if
the user only selects the desired packages and then lets the `chezmoi`
installation run automatically without any further interaction. This
way, the user only has to interact for a few minutes with the
installation itself.