feat(tmux): replace `h` and `v` by `%` and `"`
There is no need to add new key bindings when the default ones are
already good.

The key to understanding the seemingly confusing `"` and `%` is their
relative position on the US keyboard layout. `"` is positioned on the
right and `%` on the top. It is nice to think of the left-right motion
as the vertical split, and the up-down motion as the horizontal split.

Note that this commit swaps the default behaviour of `"` and `%` as it
aligns with `vim`s behaviour. See here for more information on this
behaviour: `https://github.com/tmux/tmux/issues/213`.