ci(gitleak): ignore generic api key in `gitconfig` file
"You can ignore specific findings by creating a .gitleaksignore file at
the root of your repo. In release v8.10.0 Gitleaks added a Fingerprint
value to the Gitleaks report. Each leak, or finding, has a Fingerprint
that uniquely identifies a secret. Add this fingerprint to the
.gitleaksignore file to ignore that specific secret. See Gitleaks'
.gitleaksignore for an example. Note: this feature is expirmental and is
subject to change in the future."
(https://github.com/zricethezav/gitleaks)