feat(set_wallpaper): add script to set a random wallpaper
This script can be called from anywhere in the system. This is
convenient because of this script's refactored nature and the fact that
broken wallpapers can easily be fixed by manually calling it. The
refactored nature assures the user that the wallpaper is set in the same
way it always is.