revert: "feat(nvim): add `diffview.nvim` plugin"
This reverts commit ddfe467aeb67b1e51de13d5b54c570cfdc5cc5b9.

This plugin was never really used anyway. Removing it improves startup
time by 20ms on my machine, which is on the high end of startup time per
plugin.

This nvim plugin is removed because I intend to replace it with the
`lazygit` UI for git commands.

The `neogit` plugin stays due to how easy it is to access staged files
within nvim without having to start another `lazygit` instance or do
other uncomfortably slow tricks. This use case is the number one reason
I still use the `neogit` plugin during almost every commit.