feat(set_wallpaper): add support for all image aspect ratios
The more the image and screen aspect ratios differ, the less of the
image is visible on the screen when using feh's --bg-fill flag.
Essentially this feature switches from feh's --bg-fill flag to its
--bg-max flag without causing black borders.

This script now sets a random wallpaper and uses a blurred copy of it
for the borders instead of black borders.