build(chezmoi)!: improve package installation process
Introduce a new package installation process that is designed to be more
modular and portable.

The new process automatically executes scripts and installs packages
without prompting the user to make any selections.

The new process supports installing packages either system-wide or in
the home directory depending on the values of the IS_INSTALL_PACKAGES
and IS_INSTALL_LOCALLY settings, and the availability of package
managers and sudo access. If the IS_INSTALL_PACKAGES variable is set to
true, the packages will be installed, otherwise, they will be skipped.
If IS_INSTALL_LOCALLY is false, and a supported package manager is found
with sudo access, the packages will be installed system-wide using that
package manager. If IS_INSTALL_LOCALLY is true, the packages will be
installed in the home directory using the JuNest tool if it is
available, otherwise JuNest will be installed first. If
IS_INSTALL_LOCALLY is false, no supported package manager is found, and
sudo access is not available, an error message will be printed to
standard error.

The new installation process depends on the `curl`, `git`, and `wget`
utilities for properly installing `JuNest`.

BREAKING CHANGE: The old installation process has been removed entirely.
Users will no longer be prompted to execute scripts, nor to select
packages or actions. However, the actions executed and packages
installed by the old process remain unchanged.