docs(readme): update overall structure and content
To improve structure, group all previous sections under a single
top-level `dotfiles` section.

Rename the `How to Use` section to `Installation` and update its
subsections to better reflect their purpose. `chezmoi is not installed
on the system` is now called `Installing dotfiles without an existing
chezmoi installation`, and `chezmoi is installed on the system` is now
called `Installing dotfiles with an existing chezmoi installation`.

Improve the accuracy, clarity, and detail of the `Installation` section
by updating its content. Additionally, provide information on the new
internal package installation process, including its requirements and
how to configure it.

Update the `Related` section to reflect the latest information.

Rephrase the `Contributing` section.