feat(modules/homeManager/programs/zellij): change default layout order
This layout is based on the default layout generated by the 'zellij
setup --dump-swap-layout default' command. It swaps the order of the
'vertical' and 'horizontal' layouts, such that the first layout is the
'horizontal' one.

This behaviour is preferred because Zellij's horizontal split
complements the terminal emulator's usual vertical split.

Related: https://github.com/zellij-org/zellij/discussions/2887