docs(docs/user_documentation/home_manager_configurations): improve link
Improve the 'home-configurations-full' link by replacing the static file
URL with a GitHub blob URL.

This improves the overall user experience by providing the file in its
complete context and directly viewing the file without prompting users
to download the file first, as most browsers do.

This decreases the reproducibility of the 'home-configurations-full'
link because the blob URL merely tracks the 'master' branch. However,
considering that the GitHub Pages site is updated when a 'push' event on
the 'master' branch occurs, the blob URL should be in sync with the
latest deployed documentation. Note that blob URLs may be outdated for
older documentations not pointing to latest 'master' commit.

Related: f0e3554770e544efec3cbacbf61f36d84bdcd19b